Remove existing relations with straight link
Because otherwise nested form conflics with uniqueness validations, like
validates :dish_id, uniqueness: { scope: :tag }.
The problem appear when we delete some relation, and add relation with
same uniq scope at same time.
And validation fails because it runs before database, and don't know
that we removed relation.

The problem is known in rails, e.g.
https://github.com/rails/rails/issues/18573, and atm they don't do
anything with it.